Use conditional compilation for webauthn operations
commit-id:bc26aaaf
