+++ packages/account-wasm/src/lib.rs
@@ -1,6 +1,5 @@
 mod account;
 mod errors;
 mod session;
-mod signer;
 mod types;
 mod utils;
