+++ packages/account_sdk/Cargo.toml
@@ -28,7 +28,6 @@ reqwest.workspace = true
 serde.workspace = true
 serde_cbor_2 = { version = "0.12.0-dev" }
 serde_json.workspace = true
-serde-wasm-bindgen.workspace = true
 serde_with.workspace = true
 sha2.workspace = true
 starknet.workspace = true
@@ -42,14 +41,11 @@ indexmap.workspace = true
 num-traits.workspace = true
 tsify-next = "0.5.4"
 urlencoding = "2.1.3"
-webauthn-rs-proto.workspace = true
+once_cell.workspace = true
 
 [dev-dependencies]
 hyper = { version = "0.14.27", features = ["server"] }
 rand_core = { version = "0.6", features = ["getrandom"] }
-once_cell.workspace = true
-webauthn-authenticator-rs = { workspace = true, features = ["softpasskey"] }
-webauthn-rs-core.workspace = true
 
 [build-dependencies]
 cainome.workspace = true
@@ -58,13 +54,19 @@ serde_json.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio.workspace = true
+webauthn-authenticator-rs = { workspace = true, features = ["softpasskey"] }
+webauthn-rs-core.workspace = true
+webauthn-rs-proto.workspace = true
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
+js-sys = "0.3.70"
+serde-wasm-bindgen.workspace = true
 wasm-bindgen-futures.workspace = true
 wasm-bindgen.workspace = true
 wasm-bindgen-test = "0.3.34"
-js-sys = "0.3.70"
+webauthn-rs-proto = { workspace = true, features = ["wasm"] }
 web-sys = { version = "0.3", features = [
+    "console",
     "CredentialCreationOptions",
     "CredentialRequestOptions",
     "CredentialsContainer",
