+++ packages/account_sdk/src/signers/webauthn/softpasskey_test.rs
@@ -1,26 +1,22 @@
+use crate::abigen::erc_20::Erc20;
+use crate::artifacts::Version;
+use crate::signers::webauthn::WebauthnSigner;
+use crate::signers::Signer;
+use crate::tests::account::FEE_TOKEN_ADDRESS;
+use crate::tests::runners::katana::KatanaRunner;
+
 use cainome::cairo_serde::{ContractAddress, U256};
 use starknet::{
     core::types::{BlockId, BlockTag},
     macros::felt,
 };
 
-use crate::{
-    abigen::erc_20::Erc20,
-    artifacts::Version,
-    signers::{webauthn::WebauthnSigner, Signer},
-    tests::{account::FEE_TOKEN_ADDRESS, runners::katana::KatanaRunner},
-};
-
-use super::ensure_txn;
-
 pub async fn test_verify_execute(signer: Signer) {
     let runner = KatanaRunner::load();
     let controller = runner
         .deploy_controller("username".to_owned(), signer, Version::LATEST)
         .await;
-
     let new_account = ContractAddress(felt!("0x18301129"));
-
     let contract_erc20 = Erc20::new(*FEE_TOKEN_ADDRESS, &controller);
 
     contract_erc20
@@ -30,36 +26,33 @@ pub async fn test_verify_execute(signer: Signer) {
         .await
         .expect("failed to call contract");
 
-    ensure_txn(
-        contract_erc20.transfer(
+    contract_erc20
+        .transfer(
             &new_account,
             &U256 {
                 low: 0x10_u128,
                 high: 0,
             },
-        ),
-        runner.client(),
+        )
+        .send()
+        .await
+        .unwrap();
+}
+
+#[tokio::test]
+async fn test_verify_execute_webautn() {
+    let signer = WebauthnSigner::register(
+        "cartridge.gg".to_string(),
+        "username".to_string(),
+        "challenge".as_bytes(),
     )
     .await
     .unwrap();
+
+    test_verify_execute(Signer::Webauthn(signer)).await;
 }
 
 #[tokio::test]
-async fn test_verify_execute_webauthn() {
-    let signer = Signer::Webauthn(
-        WebauthnSigner::register(
-            "cartridge.gg".to_string(),
-            "username".to_string(),
-            "challenge".as_bytes(),
-        )
-        .await
-        .unwrap(),
-    );
-
-    test_verify_execute(signer).await;
-}
-
-#[tokio::test]
-async fn test_verify_execute_starkpair() {
+async fn test_verify_execute_starknet() {
     test_verify_execute(Signer::new_starknet_random()).await;
 }
