+++ packages/account_sdk/src/tests/account/mod.rs
@@ -4,9 +4,7 @@ use starknet::macros::felt;
 
 pub(crate) mod declare;
 mod deploy;
-// mod interface;
 mod pending;
-pub(crate) mod webauthn;
 
 pub use declare::AccountDeclaration;
 pub use deploy::AccountDeployment;
