+++ packages/account_sdk/src/tests/outside_execution_test.rs
@@ -1,3 +1,9 @@
+use cainome::cairo_serde::{CairoSerde, ContractAddress, U256};
+use starknet::{
+    accounts::Account,
+    macros::{felt, selector},
+    providers::Provider,
+};
 use std::vec;
 
 use crate::{
@@ -13,17 +19,9 @@ use crate::{
     controller::Controller,
     signers::{webauthn::WebauthnSigner, Signer},
     tests::{
-        account::{webauthn::SoftPasskeySigner, FEE_TOKEN_ADDRESS},
-        ensure_txn,
-        runners::katana::KatanaRunner,
+        account::FEE_TOKEN_ADDRESS, ensure_txn, runners::katana::KatanaRunner,
         transaction_waiter::TransactionWaiter,
     },
-};
-use cainome::cairo_serde::{CairoSerde, ContractAddress, U256};
-use starknet::{
-    accounts::Account,
-    macros::{felt, selector},
-    providers::Provider,
 };
 
 pub async fn test_verify_paymaster_execute(signer: Signer, use_session: bool) {
@@ -99,7 +97,6 @@ async fn test_verify_execute_webauthn_paymaster_starknet() {
             "cartridge.gg".to_string(),
             "username".to_string(),
             "challenge".as_bytes(),
-            SoftPasskeySigner::new("https://cartridge.gg".try_into().unwrap()),
         )
         .await
         .unwrap(),
@@ -120,7 +117,6 @@ async fn test_verify_execute_webauthn_paymaster_starknet_session() {
             "cartridge.gg".to_string(),
             "username".to_string(),
             "challenge".as_bytes(),
-            SoftPasskeySigner::new("https://cartridge.gg".try_into().unwrap()),
         )
         .await
         .unwrap(),
